requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
tenacity
python-dotenv 
python-multipart
bcrypt